requirements.txt
pandas
sqlalchemy
python-dotenv
dash
plotly
numpy
fredapi
requests